implements global state, restructures code
